% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{compute,DAG.network-method}
\alias{compute,DAG.network-method}
\title{compute N samples according to the network distribution}
\usage{
\S4method{compute}{DAG.network}(object, N = 50)
}
\arguments{
\item{object}{a DAG.network object}

\item{N}{numeric. the number of samples generated according to the network}
}
\value{
a N*nNodes matrix
}
\description{
compute N samples according to the network distribution
}

