% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{computeDescriptor,D2C.Data-method}
\alias{computeDescriptor,D2C.Data-method}
\alias{computeDescriptor}
\title{computeDescriptor}
\usage{
\S4method{computeDescriptor}{D2C.Data}(object, functionDescriptor, lin = TRUE)
}
\arguments{
\item{object}{a D2C.data object}

\item{functionDescriptor}{: a descriptor function used in the D2C algorithm}

\item{lin}{: TRUE or FALSE}
}
\value{
N*nNodes matrix
}
\description{
computeDescriptor
}

