% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{D2C-class}
\alias{D2C-class}
\title{An S4 class to store the RF model trained on the basis of the descriptors of NDAG DAGs}
\description{
An S4 class to store the RF model trained on the basis of the descriptors of NDAG DAGs
}

