% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{D2C.Data-class}
\alias{D2C.Data-class}
\title{An S4 class to store the data related to a DAG}
\description{
An S4 class to store the data related to a DAG
}
\section{Slots}{

\describe{
\item{\code{network}}{a graph generated by the D2C.network class}

\item{\code{data}}{N samples computed according to the network distribution}

\item{\code{edges}}{: a set of edges we want to predict}

\item{\code{labelEdge}}{the label of edges. 1 if there is a direct connection 0 otherwise}

\item{\code{X_descriptor}}{the descriptor of the edges computed by the D2C algorithm.}
}}

