% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{DAG.network-class}
\alias{DAG.network-class}
\title{An S4 class to store DAG.network}
\arguments{
\item{network}{: object of class "graph"}
}
\description{
An S4 class to store DAG.network
}

