% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{example}
\alias{example}
\title{stored D2C object}
\description{
small D2C object for testing D2C functionalities
}
\details{
Dataset example
}
\examples{
require(RBGL)
require(gRbase)
data(example)
print(example@mod)
## Random Forest
print(dim(example@X))
## dimension of the training set
}
\keyword{data}

