% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize,D2C.Predict-method}
\alias{initialize,D2C.Predict-method}
\title{create a data structure that store a sample of a network  distribution}
\usage{
\S4method{initialize}{D2C.Predict}(.Object, data,
  lin = stop("lin must be specified"))
}
\arguments{
\item{.Object}{a D2C.predict object}

\item{data}{: matrix of samples}

\item{lin}{: logical}
}
\value{
D2C.Predict object
}
\description{
create a data structure that store a sample of a network  distribution
}

