% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predict,D2C.Predict-method}
\alias{predict,D2C.Predict-method}
\title{predict if there is a connection between node i and node j}
\usage{
\S4method{predict}{D2C.Predict}(object, i, j,
  networkType = stop("networkType must be defined"),
  functionType = stop("stop function type must be defined"), RF = NULL)
}
\arguments{
\item{object}{: a D2C.Predict object}

\item{i}{: node name}

\item{j}{: node name}

\item{networkType}{"small" , "medium" , "large'}

\item{functionType}{"linear" , "quadratic", "sigmoid"}

\item{RF}{a randomForest object}
}
\value{
list with  response and prob of the prediction
}
\description{
predict if there is a connection between node i and node j
}
\references{
Gianluca Bontempi, Maxime Flauder (2014) From dependency to causality: a machine learning approach. Under submission
}

