% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{simulatedDAG-class}
\alias{simulatedDAG-class}
\title{An S4 class to store a list of DAGs and associated observations}
\arguments{
\item{list.DAGs}{: list of stored DAGs}

\item{list.observationsDAGs}{: list of observed datasets, each sampled from the corresponding member of list.DAGs}

\item{NDAG}{: number of DAGs.}

\item{functionType}{: type of the dependency. It is of class "character" and is one of  ("linear", "quadratic","sigmoid")}

\item{seed}{: random seed}
}
\description{
An S4 class to store a list of DAGs and associated observations
}

