% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{true.net}
\alias{true.net}
\title{Adjacency matrix of the Alarm dataset}
\description{
contains the adjacency matrix of the Alarm DAG. See the vignette for an utilization of the dataset
}
\details{
Adjacency matrix of the Alarm benchmark
}
\references{
Aliferis C, Statnikov A, Tsamardinos I, Mani S, Koutsoukos X Local Causal and Markov Blanket Induction for Causal Discovery and Feature Selection for Classif ication Part II: Analysis and Extensions' by ; JMLR 2010'
}
\keyword{data}

