% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{update,simulatedDAG-method}
\alias{update,simulatedDAG-method}
\title{update of a "simulatedDAG" with a list of DAGs and associated observations}
\usage{
\S4method{update}{simulatedDAG}(object, list.DAGs, list.observationsDAGs)
}
\arguments{
\item{object}{:  simulatedDAG to be updated}

\item{list.DAGs}{: list of stored DAGs}

\item{list.observationsDAGs}{: list of observed datasets, each sampled from the corresponding member of list.DAGs}
}
\description{
update of a "simulatedDAG" with a list of DAGs and associated observations
}

