% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{updateD2C,D2C-method}
\alias{updateD2C,D2C-method}
\title{update of a "D2C" with a list of DAGs and associated observations}
\usage{
\S4method{updateD2C}{D2C}(object, sDAG, verbose = TRUE, goParallel = FALSE)
}
\arguments{
\item{object}{:  D2C to be updated}

\item{sDAG}{: simulatedDAG object to update D2C}

\item{verbose}{: TRUE or FALSE}

\item{goParallel}{: if TRUE it uses  parallelism}
}
\description{
update of a "D2C" with a list of DAGs and associated observations
}

