% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{D2C-class}
\alias{D2C-class}
\title{An S4 class to store the RF model trained on the basis of the descriptors of NDAG DAGs}
\description{
An S4 class to store the RF model trained on the basis of the descriptors of NDAG DAGs
}

