% Generated by roxygen2 (4.0.2): do not edit by hand
\name{descriptor}
\alias{descriptor}
\title{compute descriptor}
\usage{
descriptor(D, ca, ef, ns = min(4, NCOL(D) - 2), lin = FALSE, acc = TRUE,
  struct = TRUE, pq = c(0.1, 0.25, 0.5, 0.75, 0.9))
}
\arguments{
\item{D}{the data matrix}

\item{ca}{node label of the cause}

\item{ef}{node label of the effect}

\item{ns}{number of iteration in the mimr algorithm}

\item{lin}{TRUE or false functional dependency}

\item{acc}{TRUE or false use the accuracy of the regression as a descriptor}

\item{struct}{TRUE or false to use the ranking in the markov blanket as a descriptor}

\item{pq}{a vector of quantile used to compute de descriptor}
}
\description{
compute descriptor
}
\references{
Gianluca Bontempi, Maxime Flauder (2014) From dependency to causality: a machine learning approach. Under submission
}

